Privacy improvements & other minor changes (#34)
* Redo how admin roles work

* isPrivate data migration

* Add private threads and `!hide` command

* Require a number argument

* Minor newline changes

* Add getAuthUser and isDashAdmin functions

* Don't send private threads or allow access to them

* Omit private threads from appearing

* Make sure bot never ignores user

* Change how the mention system works

- Admins will no longer get mentioned if they move the thread themselves
- Admins will only get mentioned if they move the thread to a category other than the new thread category

Also, the bot will automatically make threads private/public when moved.

* Config updates

* Make thread private or public upon closure

* Require an argument at all times

* Use `req` (ty bean)

+++ src/modules/logs.js
@@ -2,7 +2,6 @@ const Eris = require("eris");
 const threadUtils = require("../threadUtils");
 const threads = require("../data/threads");
 const moment = require("moment");
-const config = require("../config");
 const utils = require("../utils");
 
 /**
@@ -14,7 +13,14 @@ module.exports = bot => {
      * @param {String} userId
      */
     async function getLogs(userId) {
-      const userThreads = await threads.getClosedThreadsByUserId(userId);
+      let userThreads = await threads.getClosedThreadsByUserId(userId);
+
+      if (! utils.isAdmin(msg.member)) {
+        userThreads = userThreads.filter((t) => ! t.isPrivate);
+      }
+
+      if (! userThreads.length) return bot.createMessage(msg.channel.id, "No logs found.");
+
       // Descending by date
       userThreads.sort((a, b) => {
         if (a.created_at > b.created_at) return -1;
@@ -27,9 +33,6 @@ module.exports = bot => {
         const formattedDate = moment.utc(thread.created_at).format("YYYY-MM-DD HH:mm [UTC]");
         return `\`${formattedDate}\` ${thread.scheduled_close_name}: <${logUrl}>`;
       }));
-
-      if (! userThreads || ! userThreads.length) return bot.createMessage(msg.channel.id, "No logs found.");
-
       const message = `**Log files for <@${userId}>:**\n${threadLines.join("\n")}`;
 
       // Send the list of logs in chunks of 15 lines per message
@@ -53,24 +56,24 @@ module.exports = bot => {
       bot.createMessage(msg.channel.id, `Deleted log files for <@!${userId}>`);
     }
 
+    let userId = thread && thread.user_id;
+
     if (args.length > 0) {
-      if (args[0] === "delete") {
+      if (args[0] === "delete" && utils.isAdmin(msg.member)) {
         const userId = utils.getUserMention(args.slice(1).join(" "));
         if (! userId) return utils.postSystemMessageWithFallback(msg.channel, thread, "Please provide a user mention or ID!");
-        if (! msg.member.roles || ! msg.member.roles.includes(config.inboxAdminRoleId)) return;
 
         return deleteLogs(userId);
       }
 
-      // User mention/id as argument
-      const userId = utils.getUserMention(args.join(" "));
-      if (! userId) return utils.postSystemMessageWithFallback(msg.channel, thread, "Please provide a user mention or ID!");
-
-      getLogs(userId);
-    } else if (thread) {
-      // Calling !logs without args in a modmail thread returns the logs of the user of that thread
-      getLogs(thread.user_id);
+      userId = utils.getUserMention(args.join(" "));
     }
+
+    if (! userId) return utils.postSystemMessageWithFallback(msg.channel, thread, "Please provide a user mention or ID!");
+
+    // Calling !logs without args in a modmail thread returns the logs of the user of that thread
+
+    return getLogs(userId);
   });
 
   threadUtils.addInboxServerCommand(bot, "loglink", async (msg, args, thread) => {
